docker implementation and project refactor
