+++ Dockerfile
@@ -1,21 +1,20 @@
-# Generated by https://smithery.ai. See: https://smithery.ai/docs/config#dockerfile
+# Use an official Python runtime as a parent image
 FROM python:3.11-alpine
 
-# Set working directory
+# Set the working directory in the container
 WORKDIR /app
 
 # Install build dependencies
 RUN apk add --no-cache gcc musl-dev linux-headers
 
-# Copy requirements and project files
-COPY pyproject.toml ./
-COPY requirements.txt ./
-COPY README.md ./
+# Copy project files into the container
+COPY pyproject.toml .
+COPY requirements.txt .
+COPY README.md .
 COPY src/ ./src/
 COPY uv.lock ./
 
-# Install pipenv system dependencies
-# Install project dependencies using pip
+# Install Python dependencies
 RUN pip install --upgrade pip \
     && pip install --ignore-installed -r requirements.txt || true
 
@@ -23,6 +22,7 @@ RUN pip install --upgrade pip \
 RUN pip install hatchling \
     && pip install .
 
-# Expose any port if needed (not strictly needed for stdio based server)
+# Expose the port the MCP server runs on
+EXPOSE 8080
 
-CMD ["python", "src/server/server.py"]
+CMD ["python", "src/server.py", "start", "--protocol", "sse", "--port", "8080"]
