+++ src/constants.py
@@ -1,0 +1,6 @@
+# Constants shared across the project
+
+CLIENT_ID = "your-client-id"
+OAUTH_HOST = "https://oauth.example.com"
+AUTH_TIMEOUT_SECONDS = 300
+ROOT_DIR = "/app"
