+++ src/init.py
@@ -3,8 +3,8 @@ import json
 import sys
 from pathlib import Path
 from typing import Optional, Literal
-from server.auth import get_authentication_token
-from server.config.app_config import AuthMethod, app_config
+from auth import get_authentication_token
+from src.config.app_config import AuthMethod, app_config
 
 # Supported client types
 ClientType = Literal["claude", "cursor", "windsurf", "copilot"]
